lambdas/process_frames/requirements.txt
pillow==10.3.0
opencv-contrib-python-headless==4.10.0.84
numpy==1.26.4